Add automatic calculation of Alejandro's Constant
I currently don't like the way the code handles the INPUT_DEADZONE and
the MOTOR_MOVEPOINT. As the EXPONENT increases the curve quickly reaches
diminishing returns to the point where EXPONENT=100 is not significantly
different than EXPONENT=5. I will look into an alternate method that
allows for more configuration of the curve.